perf(ui): share deferred activation scheduler
