+++ packages/ui/src/components/custom/deferred-external-script/index.tsx
@@ -2,8 +2,7 @@
 
 import { useEffect } from 'react'
 
-const ACTIVATION_EVENTS = ['pointerdown', 'keydown', 'touchstart'] as const
-const DEFAULT_DELAY = 5000
+import { scheduleDeferredActivation } from '@nl/ui/lib/deferred-activation'
 
 interface DeferredExternalScriptProps {
   id: string
@@ -15,62 +14,23 @@ interface DeferredExternalScriptProps {
 export default function DeferredExternalScript({
   id,
   src,
-  delay = DEFAULT_DELAY,
+  delay,
 }: DeferredExternalScriptProps): null {
   useEffect(() => {
-    let activated = false
-    let idleId: number | null = null
-    let timeoutId: number | null = null
+    const cleanup = scheduleDeferredActivation({
+      delay,
+      onActivate: () => {
+        if (document.getElementById(id)) return
 
-    const removeActivationListeners = () => {
-      for (const eventName of ACTIVATION_EVENTS) {
-        window.removeEventListener(eventName, activate)
-      }
-    }
+        const script = document.createElement('script')
+        script.id = id
+        script.src = src
+        script.async = true
+        document.head.appendChild(script)
+      },
+    })
 
-    const cancelIdleActivation = () => {
-      if (idleId !== null) {
-        window.cancelIdleCallback?.(idleId)
-        idleId = null
-      }
-      if (timeoutId !== null) {
-        window.clearTimeout(timeoutId)
-        timeoutId = null
-      }
-    }
-
-    const activate = () => {
-      if (activated) return
-      activated = true
-      removeActivationListeners()
-      cancelIdleActivation()
-
-      if (document.getElementById(id)) return
-
-      const script = document.createElement('script')
-      script.id = id
-      script.src = src
-      script.async = true
-      document.head.appendChild(script)
-    }
-
-    for (const eventName of ACTIVATION_EVENTS) {
-      window.addEventListener(eventName, activate, { once: true, passive: true })
-    }
-
-    if (window.requestIdleCallback) {
-      timeoutId = window.setTimeout(() => {
-        timeoutId = null
-        idleId = window.requestIdleCallback(activate, { timeout: 1000 })
-      }, delay)
-    } else {
-      timeoutId = window.setTimeout(activate, delay)
-    }
-
-    return () => {
-      removeActivationListeners()
-      cancelIdleActivation()
-    }
+    return cleanup
   }, [delay, id, src])
 
   return null
